Re-establish the SRT connection when restarting the pipeline
We close and reopen the SRT socket because 1) sometimes media players take a
while to recover / resync if we restart the stream over the same connection
and 2) because if we take too long, the SRT connection will timeout and
we'll just get an SRT error as soon the pipeline succesfully restarts.

1) relies on any SRT relay servers to close the connection to the video
player when the ingest connection is closed, and also on the video player to
reconnect. This results in reliable, speedy recovery when using OBS and the
Belabox Cloud SRT relay servers