Dynamic bitrate: cap BS threshold depending on current throughput
This attempts to address a corner case on networks which throttle
network speed by immediately dropping packets rather than queueing
them. When such a network is congested, the RTT remains relatively
constant while packets are dropped. In the SRT stats, this is
reflected by the send buffer size (SRTO_SNDDATA) rapidly increasing.

However, SRTO_SNDDATA has an upper bound depending on the current
throughput and the SRT latency settings, as SRT discards packets
which are too old for timely delivery. In some conditions, it was
possible for the BS bitrate reduction thresholds in belacoder to
increase beyond this upper bound before the bitrate was reduced
to the minimum level and while the network was still congested and
dropping packets.

This patch was only tested using a simulation of such a network, as
none of the ISPs I use works in this way. Visible Wireless users
are highly encouraged to apply this patch and do a test stream with
the bitrate overlay enabled