Set PTS to 0 instead of outputting multiple frames with the same PTS
When the input resolution to a camlink 4k connected to a jetson is
changed, we get a bunch of frames at a shorter interval than normal
for the selected framerate, then the v4l2src element stalls. See
3ea04da5b for details and the workaround using stall detection

However, when combining the ptsfixup element with stall detection,
the duplicated PTS on consecutive frames sometimes made the
pipeline to appear to continue running for some time, breaking
stall detection

This patch sets a PTS of 0 when a frame arrives with a PTS that's
too old in relation to our 'fixedup' PTS, which avoids the issue
on our Jetson pipelines.

We could drop the frame entirely, but at this point we're just
building workarounds for our workarounds. The camlink stall doesn't
happen on more up-to-date systems, so hopefully we'll be able to
backport a proper fix later on